Sync chat history per-message, harden DB, unify rubric
- Persist each user/assistant turn via /api/session/start + /api/session/append; resume in-progress runs from /api/session/active on chat hydrate.
- Wrap session writes in transactions, raise pool to 5 with idle/connect timeouts, and add (player_id, character_id, last_message_at) index for resume queries.
- Admin reset now TRUNCATEs players + runs + messages; dashboard messaging and response payload report deleted players too.
- Replace refusal access keyword from "차은우" to "국세청".
- Remove admin persona-editing surface (PersonaSettings UI, /api/admin/personas, /api/admin/avatar-options, upsertPersonaConfig).
- Stop tiny 1–2 point affection swings: clarify trigger vs rubric scales in the system prompt, infer 0 (not 1) for neutral status, and lift small absolute boosts so |1|/|2| step to ±5/±6.
- Unify rubric to integer 10-point totals across all characters and require integer points in the schema.

+++ lib/characters.ts
@@ -11,7 +11,7 @@ const ruleSchema = z.object({
 
 const evaluationRubricItemSchema = z.object({
   label: z.string().min(1),
-  points: z.number().positive(),
+  points: z.number().int().positive(),
   criteria: z.string().min(1),
 });
